removed required:true property on kilifi for time being.
Warning: this breaks muvuku validation.

required:true causes the validations in sentinel/kujua to not run if the
field is missing or not parsed correctly.  this is a problem in how the
the transitions.  the sentinel validation should run independently on
all records and the registration transition should run after validation.

this was causing a validation problem so had to make a temporary fix. if
these forms move to muvuku which is likely in the future the muvuku
validations we'll need to fix things to muvuku validations and sentinel
validations can both run successfully.